// src/app/models/appointments/reschedule-appointment.ts
export class RescheduleAppointment {
    appointment_id: string;
    appointment_no: number;
    appointment_from_time: string;
    appointment_to_time: string;
    appointment_date: string;
    schedule_id: string;
    hospital_id: string;
    doctor_id: string;
    contact_id: string;
    patient_name: string;
    birth_date: string;
    medical_record_number: string;
    patient_phone_number: string;
    queue_number: string;
    note: string;
    is_waiting_list: boolean;
    appointment_status_id: number;
    created_by: string;
    created_date: string;
    created_from: string;
    modified_by: string;
    modified_date: string;
    modified_from: string;
  }